requirements.txt
requests
pyserial
vosk
pyaudio
pvporcupine
simpleaudio
